Replace progress with qprogress, fix import in threading
